SINAN disease returns circular import
