+++ pysus/online_data/CNES.py
@@ -1,5 +1,6 @@
 from typing import Union
-from pysus.online_data import FTP_Downloader, CACHEPATH
+
+from pysus.online_data import CACHEPATH, FTP_Downloader
 
 group_dict = {
     "LT": ["Leitos - A partir de Out/2005", 10, 2005],
@@ -19,10 +20,14 @@ group_dict = {
 
 
 def download(
-        group: str, states: Union[str, list], years: Union[str, list, int], months: Union[str, list, int], data_dir: str=CACHEPATH
+    group: str,
+    states: Union[str, list],
+    years: Union[str, list, int],
+    months: Union[str, list, int],
+    data_dir: str = CACHEPATH,
 ) -> list:
     """
-    Download CNES records for group, state, year and month and returns a 
+    Download CNES records for group, state, year and month and returns a
     list of local parquet files
     :param group:
         LT – Leitos - A partir de Out/2005
@@ -42,7 +47,7 @@ def download(
     :param states: 2 letter state code, can be a list of UFs
     :param years: 4 digit integer, can be a list of years
     """
-    return FTP_Downloader('CNES').download(
+    return FTP_Downloader("CNES").download(
         CNES_group=group,
         UFs=states,
         years=years,
