+++ pysus/online_data/PNI.py
@@ -2,10 +2,15 @@
 Download data from the national immunization program
 """
 from typing import Union
-from pysus.online_data import FTP_Downloader, FTP_Inspect,CACHEPATH
+
+from pysus.online_data import CACHEPATH, FTP_Downloader, FTP_Inspect
 
 
-def download(states: Union[str, list], years: Union[str, list, int], data_dir: str=CACHEPATH) -> list:
+def download(
+    states: Union[str, list],
+    years: Union[str, list, int],
+    data_dir: str = CACHEPATH,
+) -> list:
     """
     Download imunization records for a given States and years.
     :param state: uf two letter code, can be a list
@@ -13,11 +18,8 @@ def download(states: Union[str, list], years: Union[str, list, int], data_dir: s
     :param data_dir: directory where data will be downloaded
     :return: list of downloaded parquet paths
     """
-    return FTP_Downloader('PNI').download(
-        PNI_group='CPNI',
-        UFs=states,
-        years=years,
-        local_dir=data_dir
+    return FTP_Downloader("PNI").download(
+        PNI_group="CPNI", UFs=states, years=years, local_dir=data_dir
     )
 
 
@@ -27,11 +29,8 @@ def get_available_years(state):
     :param state: uf code
     :return: list of strings (filenames)
     """
-    return FTP_Inspect('PNI').list_available_years(
-        UF=state,
-        PNI_group='CPNI'
-    )
+    return FTP_Inspect("PNI").list_available_years(UF=state, PNI_group="CPNI")
 
 
 def available_docs():
-    return FTP_Inspect('PNI').list_all(PNI_group='CPNI')
+    return FTP_Inspect("PNI").list_all(PNI_group="CPNI")
