+++ pysus/online_data/SIM.py
@@ -1,21 +1,24 @@
-u"""
+"""
 Download Mortality records from SIM Datasus
 Created on 12/12/18
 by fccoelho
 license: GPL V3 or Later
 """
 import os
-import pandas as pd
-
-from dbfread import DBF
-from typing import Union
-from loguru import logger
 from ftplib import FTP, error_perm
+from typing import Union
 
+import pandas as pd
+from dbfread import DBF
+from loguru import logger
 from pysus.online_data import CACHEPATH, FTP_Downloader
 
 
-def download(states: Union[str, list], years: Union[str, list, int], data_dir: str=CACHEPATH):
+def download(
+    states: Union[str, list],
+    years: Union[str, list, int],
+    data_dir: str = CACHEPATH,
+):
     """
     Downloads data directly from Datasus ftp server
     :param states: two-letter state identifier: MG == Minas Gerais
@@ -23,10 +26,8 @@ def download(states: Union[str, list], years: Union[str, list, int], data_dir: s
     :param years: 4 digit integer, can be a list
     :return: a list of downloaded parquet paths
     """
-    return FTP_Downloader('SIM').download(
-        UFs=states,
-        years=years,
-        local_dir=data_dir
+    return FTP_Downloader("SIM").download(
+        UFs=states, years=years, local_dir=data_dir
     )
 
 
@@ -38,19 +39,25 @@ def get_CID10_chapters_table(cache=True):
     """
     ftp = FTP("ftp.datasus.gov.br")
     ftp.login()
-    logger.debug(f"Stablishing connection with ftp.datasus.gov.br.\n{ftp.welcome}")
+    logger.debug(
+        f"Stablishing connection with ftp.datasus.gov.br.\n{ftp.welcome}"
+    )
     ftp.cwd("/dissemin/publicos/SIM/CID10/TABELAS")
-    logger.debug("Changing FTP work dir to: /dissemin/publicos/SIM/CID10/TABELAS")
+    logger.debug(
+        "Changing FTP work dir to: /dissemin/publicos/SIM/CID10/TABELAS"
+    )
 
     fname = "CIDCAP10.DBF"
-    cachefile = os.path.join(CACHEPATH, "SIM_" + fname.split(".")[0] + "_.parquet")
-    
+    cachefile = os.path.join(
+        CACHEPATH, "SIM_" + fname.split(".")[0] + "_.parquet"
+    )
+
     if os.path.exists(cachefile):
         logger.info(f"Local parquet file found at {cachefile}")
         df = pd.read_parquet(cachefile)
 
         return df
-        
+
     try:
         ftp.retrbinary("RETR {}".format(fname), open(fname, "wb").write)
 
@@ -78,12 +85,18 @@ def get_CID10_table(cache=True):
     """
     ftp = FTP("ftp.datasus.gov.br")
     ftp.login()
-    logger.debug(f"Stablishing connection with ftp.datasus.gov.br.\n{ftp.welcome}")
+    logger.debug(
+        f"Stablishing connection with ftp.datasus.gov.br.\n{ftp.welcome}"
+    )
     ftp.cwd("/dissemin/publicos/SIM/CID10/TABELAS")
-    logger.debug("Changing FTP work dir to: /dissemin/publicos/SIM/CID10/TABELAS")
+    logger.debug(
+        "Changing FTP work dir to: /dissemin/publicos/SIM/CID10/TABELAS"
+    )
 
     fname = "CID10.DBF"
-    cachefile = os.path.join(CACHEPATH, "SIM_" + fname.split(".")[0] + "_.parquet")
+    cachefile = os.path.join(
+        CACHEPATH, "SIM_" + fname.split(".")[0] + "_.parquet"
+    )
 
     if os.path.exists(cachefile):
         logger.info(f"Local parquet file found at {cachefile}")
@@ -118,12 +131,18 @@ def get_CID9_table(cache=True):
     """
     ftp = FTP("ftp.datasus.gov.br")
     ftp.login()
-    logger.debug(f"Stablishing connection with ftp.datasus.gov.br.\n{ftp.welcome}")
+    logger.debug(
+        f"Stablishing connection with ftp.datasus.gov.br.\n{ftp.welcome}"
+    )
     ftp.cwd("/dissemin/publicos/SIM/CID9/TABELAS")
-    logger.debug("Changing FTP work dir to: /dissemin/publicos/SIM/CID9/TABELAS")
+    logger.debug(
+        "Changing FTP work dir to: /dissemin/publicos/SIM/CID9/TABELAS"
+    )
 
     fname = "CID9.DBF"
-    cachefile = os.path.join(CACHEPATH, "SIM_" + fname.split(".")[0] + "_.parquet")
+    cachefile = os.path.join(
+        CACHEPATH, "SIM_" + fname.split(".")[0] + "_.parquet"
+    )
 
     if os.path.exists(cachefile):
         logger.info(f"Local parquet file found at {cachefile}")
@@ -158,19 +177,25 @@ def get_municipios(cache=True):
     """
     ftp = FTP("ftp.datasus.gov.br")
     ftp.login()
-    logger.debug(f"Stablishing connection with ftp.datasus.gov.br.\n{ftp.welcome}")
+    logger.debug(
+        f"Stablishing connection with ftp.datasus.gov.br.\n{ftp.welcome}"
+    )
     ftp.cwd("/dissemin/publicos/SIM/CID10/TABELAS")
-    logger.debug("Changing FTP work dir to: /dissemin/publicos/SIM/CID10/TABELAS")
+    logger.debug(
+        "Changing FTP work dir to: /dissemin/publicos/SIM/CID10/TABELAS"
+    )
 
     fname = "CADMUN.DBF"
-    cachefile = os.path.join(CACHEPATH, "SIM_" + fname.split(".")[0] + "_.parquet")
+    cachefile = os.path.join(
+        CACHEPATH, "SIM_" + fname.split(".")[0] + "_.parquet"
+    )
 
     if os.path.exists(cachefile):
         logger.info(f"Local parquet file found at {cachefile}")
         df = pd.read_parquet(cachefile)
 
         return df
-        
+
     try:
         ftp.retrbinary("RETR {}".format(fname), open(fname, "wb").write)
 
@@ -198,11 +223,17 @@ def get_ocupations(cache=True):
     """
     ftp = FTP("ftp.datasus.gov.br")
     ftp.login()
-    logger.debug(f"Stablishing connection with ftp.datasus.gov.br.\n{ftp.welcome}")
+    logger.debug(
+        f"Stablishing connection with ftp.datasus.gov.br.\n{ftp.welcome}"
+    )
     ftp.cwd("/dissemin/publicos/SIM/CID10/TABELAS")
-    logger.debug("Changing FTP work dir to: /dissemin/publicos/SIM/CID10/TABELAS")
+    logger.debug(
+        "Changing FTP work dir to: /dissemin/publicos/SIM/CID10/TABELAS"
+    )
     fname = "TABOCUP.DBF"
-    cachefile = os.path.join(CACHEPATH, "SIM_" + fname.split(".")[0] + "_.parquet")
+    cachefile = os.path.join(
+        CACHEPATH, "SIM_" + fname.split(".")[0] + "_.parquet"
+    )
 
     if os.path.exists(cachefile):
         logger.info(f"Local parquet file found at {cachefile}")
